# public/api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "UniHER API",
    "description": "API da plataforma UniHER - Saúde e bem-estar corporativo feminino",
    "version": "1.0.0",
    "contact": {
      "name": "UniHER",
      "url": "https://uniher.com.br"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Servidor principal"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Token JWT obtido via login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "Cookie httpOnly com token JWT"
      }
    },
...
  },
  "paths": {
...
    "/rh/objectives": {
      "get": {
        "tags": ["RH - Objetivos"],
        "summary": "Listar objetivos (RH)",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "200": { "description": "Lista de objetivos" }
        }
      },
      "post": {
        "tags": ["RH - Objetivos"],
        "summary": "Criar objetivo",
        "security": [{ "cookieAuth": [] }],
        "responses": {
          "201": { "description": "Objetivo criado" }
        }
      }
    },
...
  },
...
}